requirements.txt
cvzone==1.5.6
mediapipe==0.8.10
paho-mqtt==1.6.1
